+++ junit5/src/main/java/com/google/testing/junit/testparameterinjector/junit5/ParameterValueParsing.java
@@ -17,10 +17,12 @@ package com.google.testing.junit.testparameterinjector.junit5;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
+import static com.google.common.collect.Iterables.getOnlyElement;
 
 import com.google.common.base.CharMatcher;
 import com.google.common.base.Function;
 import com.google.common.base.Optional;
+import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
 import com.google.common.primitives.Primitives;
 import com.google.common.primitives.UnsignedLong;
@@ -31,10 +33,12 @@ import java.lang.reflect.ParameterizedType;
 import java.math.BigInteger;
 import java.nio.charset.Charset;
 import java.util.Arrays;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Set;
 import javax.annotation.Nullable;
 import org.yaml.snakeyaml.LoaderOptions;
 import org.yaml.snakeyaml.Yaml;
@@ -131,6 +135,11 @@ final class ParameterValueParsing {
     yamlValueTransformer
         .ifJavaType(Enum.class)
         .supportParsedType(
+            Boolean.class,
+            bool ->
+                ParameterValueParsing.parseEnumIfUnambiguousYamlBoolean(
+                    bool, javaType.getRawType()))
+        .supportParsedType(
             String.class, str -> ParameterValueParsing.parseEnum(str, javaType.getRawType()));
 
     yamlValueTransformer
@@ -164,6 +173,42 @@ final class ParameterValueParsing {
         .supportParsedType(Map.class, map -> parseYamlMapToJavaMap(map, javaType));
 
     return yamlValueTransformer.transformedJavaValue();
+  }
+
+  private static Enum<?> parseEnumIfUnambiguousYamlBoolean(boolean yamlValue, Class<?> enumType) {
+    Set<String> negativeYamlStrings =
+        ImmutableSet.of("false", "False", "FALSE", "n", "N", "no", "No", "NO", "off", "Off", "OFF");
+    Set<String> positiveYamlStrings =
+        ImmutableSet.of("on", "On", "ON", "true", "True", "TRUE", "y", "Y", "yes", "Yes", "YES");
+
+    // This is the list of YAML strings that a user could have used to define this boolean. Since
+    // the user probably didn't intend a boolean but an enum (since we're expecting an enum), one of
+    // these strings may (unambiguously) match one of the enum values.
+    Set<String> yamlStringCandidates = yamlValue ? positiveYamlStrings : negativeYamlStrings;
+
+    Set<Enum<?>> matches = new HashSet<>();
+    for (Object enumValueObject : enumType.getEnumConstants()) {
+      Enum<?> enumValue = (Enum<?>) enumValueObject;
+      if (yamlStringCandidates.contains(enumValue.name())) {
+        matches.add(enumValue);
+      }
+    }
+
+    checkArgument(
+        !matches.isEmpty(),
+        "Cannot cast a boolean (%s) to an enum of type %s.",
+        yamlValue,
+        enumType.getSimpleName());
+    checkArgument(
+        matches.size() == 1,
+        "Cannot cast a boolean (%s) to an enum of type %s. It is likely that the YAML parser is"
+            + " 'wrongly' parsing one of these values as boolean: %s. You can solve this by putting"
+            + " quotes around the YAML value, forcing the YAML parser to parse a String, which can"
+            + " then be converted to the enum.",
+        yamlValue,
+        enumType.getSimpleName(),
+        matches);
+    return getOnlyElement(matches);
   }
 
   private static Map<?, ?> parseYamlMapToJavaMap(Map<?, ?> map, TypeToken<?> javaType) {